增加获取根文件夹、创建文件夹接口，完善文件夹删除功能，修改一些数据类的变量类型，完善一些可能会发生逻辑问题的地方

+++ Main/src/main/java/com/clouddrive/main/rocketmq/UploadOKMessageConsumer.java
@@ -34,7 +34,7 @@ public class UploadOKMessageConsumer implements RocketMQListener<String> {
         });
         String hash = map.get("hash").toString();
         String fileId = map.get("hash").toString();
-        Integer size = Integer.valueOf(map.get("size").toString());
+        Long size = Long.valueOf(map.get("size").toString());
 
         //读取之前保存在Redis里面的文件相关数据
         String fileLocalData = redisUtil.getString("uploadData:" + hash + ":" + size);
